flask_template/templates/simple/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
gunicorn

click
shortuuid
pycrypto
pymysql
